refactor: rename ghost for consistency and group in ghostImpl folder
