+++ source/ghostImpl/organgeGhost.py
